requirements.txt
# LangGraph & LangChain Core
langgraph>=0.2.0
langgraph-cli>=0.1.0
langgraph-checkpoint-sqlite>=0.2.0
langgraph-store>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Vector Database & RAG
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Memory & Persistence (Optional - Production)
pymongo>=4.0.0

# LM Studio 연동 (OpenAI API 호환)
openai>=1.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
pydantic>=2.0.0
typing-extensions>=4.5.0

# Monitoring (Optional)
langsmith>=0.1.0

# UI (Optional)
chainlit>=1.0.0
streamlit>=1.30.0

# Development
pytest>=7.4.0
black>=23.0.0
mypy>=1.5.0
